refactor(legend): AqiLegend 收編進 LEGEND_REGISTRY
AqiLegend 原本由 App.tsx 直接掛在畫面右下角，繞過 LEGEND_REGISTRY，
使得 layerConsistency 看不到它、把 aqiImagery / aqiStations 當成「合法
無圖例」凍進 baseline —— 這正是圖例漂移的根因之一（測試量不到的東西
就會慢慢消失）。

- components/AqiLegend.tsx 移進 LegendPanel.tsx 成為 sub-component 並刪檔
- 註冊 { keys: ["aqiImagery", "aqiStations"] }，位置接在 MicroSensorLegend 後
- App.tsx 移除手掛與 import；AqiProductSwitcher 是控制項，留在原處，
  條件從 (aqiImagery || aqiStations) 收斂成 aqiImagery（它本來就只在
  aqiImagery 開啟時顯示，外層 or 是為了包 AqiLegend 才存在）
- 未被使用的 caption prop 一併移除

色階內容不變：仍是 AQI_LEVELS 六級 + 同一組刻度 + 同一個 loading 狀態；
只拔掉自己的面板外殼（背景/模糊/圓角/220px 固定寬）改吃 LegendPanel 容器。
位置與行為必然變動：圖例改到 LEGEND 面板內，面板預設收合。

同時修掉 layerConsistency.test.ts 與事實不符的豁免註解（原註解只講
aqiMicroSensors，沒解釋 aqiImagery / aqiStations 為何在 baseline），
並補上 cwaCloudImagery / cwaRadarImagery 的真正理由（上游直接給已上色的
PNG，前端沒有色階可派生）。
